requirements.txt
numpy.py
math.py
operator.py